The initial part of the PR fixes #6261 First we attempt to fix this during upgrades, but also allow the admin to randomize the auth_secret during upgrade.
Second attempt is during normal usage. We log messages indicating it is missing but only once every 15 minutes. For admins we issue a security warning as well.

As a fallback, we set secret_auth to a static value. This weakens security, but at least allows logins to occur normally.

This also fixes some editor handling issues in Subs-Admin.php with regards to ?>
As well we have updated updateDbLastError to return a bool on success/failure and optionally updates the $db_last_error into the current context, allowing if needed for other parts of the system or 3rd party APIs to recognize we have an error that was written in the current context.

The second part of this PR is fixing multiple upgrader issues I found while making the first fix.
First:
PHP Notice: Undefined index: lang_character_set in /upgrade.php on line 687
PHP Stack trace:
PHP 1. {main}() /upgrade.php:0
PHP 2. loadEssentialData() /upgrade.php:188

We resolve this by loading the index language as well. We also give a CLI friendly error message if we can't load a language.

Second:
If we specify --path=, this causes some failures in upgrade.php. I've corrected them

$upgrade_path is defined in the global context, so we bring it into functions where it is needed and replace usages of dirname(FILE) otherwise. $upgrade_path defaults to this but CLI allows us to override it.
Also fixes it for language path error message.
Third:
If Settings.php or Settings_bak.php can't be written, we fail. This should be writable, especially for CLI case as it should be a shell account with write access to these files, even if the web user does not have write access.

Finally:
Updated the CLI options to include rebuilding settings, resetting auth and allowing stat collection. Updated the help topic as well.