Avoid pointer cursor for unresized images
If an image has been included in a message with the
img bbc code, but without specifying width or height,
the cursor changes to a pointer when hovering,
indicating that an action is attached when clicking.
But there is no action, since the image is not resized.
If the width or height is specified, the image is
toggled to full size when clicked, so there should
be a cursor pointer.
Catch img tags without width or height first to
display without the resize class.

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com