Fix problems when privacy policy is empty
If the setting for require new uesers to accept the privacy policy
is enabled, but the policy is empty, a critical error is posted
to the error log. The string was however not defined.
Define the string and also fix missing agreement warning string
by changing lang file.
Add a warning about this in the admin page, to align with missing
agreement.

Fixes #6416

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com