Set user id after registration
During the registration of a new user, custom profile fields
could be configured to be set during the registration.
When changed these custom profile field edits are added to the
profile edits log. But since the user id is not set at this time,
the changed by user ended up as 0.
Solve this by setting the user id after it has been created.

Fixes #6237

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com