Correct calculation if user can see warning
The calculation if the user can see a user warning on the
profile summary was wrong. Leading to that users with
permission to see only their own warning status, could not
do so.
The check for if the warning system is turned on, used int
to string comparison, leading to it never evaluation to true.
This was however only checked if the user viewed their own
profile.
Corrected both these issues.

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com