fix: Localize code headers in editor (#6553)
This is the missing half of #6528.I didn't want to set a global variable like some of the other codes do, and the global namespace is already too crowded. I cannot use the logical method of sceditor.locale because the bbcode format exposes the bbcode object, not the actual editor. So I added a new  option to `parserOptions` called `txtVars`.

Also adds the browser cache variable to the editor's `iframe` stylesheet.