Fix reactivation of hooks
If a hooks was disabled, it could not get reactivated again.
This was because an ! was added to the hook setting to
indicate that it is disabled. The function name with the !
was then used when trying to enable the hook.
Solved this by using the function name without !.

Also reload the page after deletion of hooks.

Fixes #5444

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com