Move max attachments logic to javascript
The logic of how many attachments remaining that could
be uploaded was handled in PHP. This resulted in that if
the max amount of attachments was attached to a post.
If modifying it, you could not remove an attachment and
add another one, without saving it in between.
Move this logic into javascript, to solve this issue.
Also clamped the remaining attatchment counter to 0.

Fixes #6582

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com