Allow CORS requests in SMF
This changes checkSession to ignore the HTTP_HOST verification if we have a valid CORS domain found, as HTTP_ORIGIN is trusted: https://stackoverflow.com/questions/4566378/how-secure-is-http-origin.  As such if we have found a valid CORS, we don't need to check the HTTP_HOST because HTTP_REFERER will never validate otherwise.

This also changes send_http_status to send the frame and cors policy headers.  In events where we used send_http_status, sometimes we issued ob_end_clean and that causes all previous headers to be removed.  This ensures we have the headers still to send.  Additionally the 204 status code is now valid default.

This also modifies the index action to move both the Scheduled Task and check that the request method is not an options call.  This is performed by CORS requests and we should avoid doing any extra work other than indicating we have a valid headers for the CORS to accept requests.

We also added a check to the Scheduled Task to ensure during an options request method, we just bail out.  This is performed in addition to the index.php requests incase an external call from SSI.php has called this.