Correct previous and next month in calendar
Sometimes if selecting a date in the beginning or end of
the month, the previous and next month was incorrectly
calculated.
This was because we added or subtracted "1 month" to the
current date. PHP treats 1 month as 31 days. So if the current
month has less than 31 days the calculation is wrong.
Insted use "first day of next month" and
"first day of previous month".

Signed-off-by: Oscar Rydhé oscar.rydhe@gmail.com