Support negated regex tests for bbcodes
This is the polar opposite of`test` where it will skip a bbcode if there is a match. 'negative_test` searches the same string as `test`, but doesn't add an anchor.

This is my fix for #6654—a negative lookahead may be possible here but I cannot get it to work. I doubt that regular expressions were made to do that.

Signed-off-by: John Rayes <live627@gmail.com>