Set forum default timezone at registration
* If the timezone field was not shown on registration
new users would get the UTC timezone set by default,
instead of the forum default timezone.
* And if the default timezone was not provided by the
caller of RegisterMember() UTC was also used. Move the
fetching of the default timezone from the caller to
the function.
* Also fixed a small error where timezone selection
would not stay if the registration form was reloaded,
e.g. if an error occured, if the forum default timezone
was selected.

Fixes #6747

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>