Show attachment directory as text and not json
If only one attachment directory exists, which is the case in a
forum newly installed or upgraded from 2.0, the attachment
directory is shown as a json field in the Attachments Settings.
Changed to showing it as text, like is done with the Current
Attachment directory if more than one attachment dir exists.

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>