Set default board profile when profile set to inherit
If creating a new board with the profile set as
"Inherit from parent board" an error will occur.
The board will be created, but without most of the
set options. Fixed this by setting the default profile
if inherit is selected. After the board is created
a check is done to find the parent board profile
setting and copy it.

Fixes #6798

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>