Fix legacy attachments upgrade
During the legacy attachments upgrade, batches of 100
rows that do not have attachment_type of 1, are queried.
But in the prosessing of these rows, attachment_type is
sometimes change to 1. When this happens, rows will be
skipped and those attachments not processed. This can
lead to file not found errors. Fixed this problem by
decreasing the index when changing attachment_type.

Fixes #6181
Fixes #6810
Closes #6177

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>