Allow sorting the member list by custom fields after search (#6834)
If searching the memberlist by a non custom profile field,
and then sorting the list by a custom profile field result
in an sql error. This is beacuse the custom field is
not joined. Fixed this by adding the custom field to the
query (if not already part since included by the search
fields), just as is done in the display member list case.

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>