Make images responsive if max width is set (#6774)
If the attachments max width is set the max-width
was set to a px value. This would override the responsive
default value of 100% and would make the images non responsive.
Instead calculate a minimum value of 100% and the selected
max width. This will make the image responsive up to the
selected max value.

Fixes #6772

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>