Fixes some upgrader issues
Fixes #7075 by adding the needed variable into the headers
Fixes #7070 by using the select method instead.  If this introduces issues because we no longer check the client we can reintroduce that.
This will conflict with #7072 if it merges first.  I changed the eval call to a closure instead.  Easier to read and Syntax highlighting works in editors.
I also added a sanity check to ensure the function exists.