Improve the hooks list
-Greatly reduces the number of calls to `file_exists()` because it's kinda slow on Windows
- Fix sorting
- Functions with the same name but have different namespaces, classes, are now resolved correctly

The previous name resolution algorithm saws tightly wound in several loops (`num_files * nun_hooks *, num_functions`) and therefore could note scale at all. This is now changed so that all filesystem operations are outside of the main loops.

Filesystem operations on Windows can be a bit slow,    (NTFS) even with a SSD. So build a list of files from `$sourcedir` and use that to check for file existence before resorting to reading the disk.

Signed-off-by: John Rayes <live627@gmail.com>