Fix the activating link in profile
If a user was not yet activated a warning and a link will
be shown in the users profile. However the link would go
to the ACP approval tab, and not the activation tab. This
is confusing a lot of people since the link used to activate
the users directly in 2.0. Fixed so that the link would go
to the activating tab when the users need to be activated and
to the approval tab when approvals are needed.
Also removed the warning when users want to be deleted, as
the action will not happen on clicking the link.

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>