Use proper error message for missing verification questions
If verification questions are configured to be used at
registration, but no questions exists for the current
language, the user is presented with the error message:
"You are not allowed to access this section". This is very
confusing, both for the user and the admin. Display a proper
error message and also log it to the error log.

Signed-off-by: Oscar Rydhé <oscar.rydhe@gmail.com>