Send our error message to our xhr handlers
Fixes #7761
A new header x-SMF-errormsg is now sent.  This can be received by xhr on the client side to be displayed.

Additionally fixes a PHP 8.2 error with $_sa being undefined in the Likes class
Ensures that $txt[$error] exists before we try to use it during a non fatal error