SMTP Mail Queueing improvements
Fixes #7787

Trying to keep inside the database schema, i'm abusing the priority level to slowly raise it to the max (127 due to signed tinyint).  It does this over a period of 3 days (which can be changed) by calculating the difference between the current time and the time the email was sent.