Fix PHP error from a failed database connection
PHP 8.1+ will generate the following error upon a failed database connection:
```
PHP Fatal error:  Uncaught TypeError: mysqli_error(): Argument #1 ($mysql) must be of type mysqli, null given in /Sources/Errors.php:464
```

To fix this, I built a wrapper to check to ensure that the we don't have a null argument going into the related function.  Instead if its null and the primary connection is null, it returns a empty string, which is what the code seems to expect.