Implements SMF\Actions\Profile\GroupMembership
In addition to refactoring, this also fixes some inconsistency in the logic regarding switching primary groups. Specifically, the old version would not offer members the ability to *switch* primary groups if the current primary was private or protected, when what was actually intended was to not offer them the ability to *leave* private or protected groups. This version continues to prevent leaving private and protected groups, but it now correctly allows switching primary groups.
Signed-off-by: Jon Stovell <jonstovell@gmail.com>