Kills off sha1_smf
This attempt at a polyfill for sha1() was introduced during the betas for SMF 1.1 in an attempt to provide support for SHA-1 hashing on PHP 4.1.x-4.2.x. If mhash() it was available it would use that, but otherwise it fell back to this userland algorithm. It never worked correctly, at least on Linux and Unix servers, which was acknowledged in the comments in Login2. But the fact that it didn't work correctly was the precisely reason why it was retained. Specifically, if someone's password had been saved in a forum running an SMF 1.1 beta on PHP 4.1.x-4.2.x on a Linux server, it would have been saved using that broken hash. So to enable on-the-fly re-encryption in the "other passwords" section of Login2 after an SMF and/or PHP upgrade, the broken userland algorithm was retained in Subs-Compat so that the saved value could be checked against it.

But here's the thing: the output produced by this userland algorithm changed in PHP 5.3, meaning that the value it returned would not have matched what as originally generated on PHP 4 anyway. And then the output changed again in PHP 5.5, again in PHP 7.2, and again in PHP 8.1. In other words, this code that was retained for the sole purpose of allowing on-the-fly password re-encryption under certain rare circumstances hasn't worked for the past 15 years. And no one even noticed. In light of that, I think it is more than fair to say we don't need it any more.

Signed-off-by: Jon Stovell <jonstovell@gmail.com>