[3.0] Overhaul User IP handling
Fixes #7684

This PR addresses the above issue but also isolates the handling of the IP determination into a dedicated function as well as removing all places where we are changing the $_SERVER['REMOTE_ADDR'] or setting $_SERVER['BAN_CHECK_IP'], except if we are in compatibility mode.  The later purpose is to have SMF stop modifying variables on the server, leaving them as was and cleaning them to its standards and using that when needed.

This fixes the PR by flipping how we have done the checks.  In the past we have determined the user being passed through the reverse proxy and setting that users real IP into BAN_CHECK_IP, leaving REMOTE_ADDR as the proxy.  We then proceeded to check and use most of the time REMOTE_ADDR.  By flipping this, we get the users real IP and use that and use the alternative IP as the original requesting IP (reverse proxy).