[3.0] Improve load average detection
This improves the load average detection by moving the logic into Sapi.
A new function is added to simplify calls to checking if the server is overloaded
Additionally it attempts to make use of the built in sys_getloadavg(), with checks incase a server admin has disabled it.

This will Fix #7397, although that is marked as a 2.1 bug.

A note about linux was added and how load is calculated.  We could run `nproc` and divide that by the load.  But Linux admins are used to the logic and it would just confuse more server admins than help.

A note about Windows support was included.  I attempted to check the speed of commands, but they where slow.  On my i9-10850k running at idle, I ran those commands through Measure-Command powershell function and they would take 100+ milliseconds to complete.  Not ideal to add 0.1+ load time just to calculate the server load.  Wmic is known to be slow to process commands.