[3.0] Change load average to percent based results
Still returns a float, however its now based on cpu load / processors to give you the percent of the cpu its based on.
I removed the mac os restrictions, as I don't see why we would otherwise be preventing it from working there.

I adjusted the defaults, but we may want to discuss them.
I didn't add a upgrader for this.  I don't feel we should change it.  If we do anything, I would suggest just turning it off and letting the admin reset things.