[3.0] Implement File System handler
This introduces a file system handler for package manager, bringing in FTPS and SFTP support to our package manager operations.

FTPS is handled with PHP ftp library and SFTP is handled with Phpseclib (MIT License).
This simplifies and centralizes the logic for determining how to perform file system operations by using the root as a substitution for the boarddir.

The test page has been changed from xml supported to json and most of the logic moved into admin.js.

The settings page now operates as config vars.  Not sure why it wasn't ever this way.

The legacy FtpConnection was moved to basically be a wrapper for the new file system caller.  I didn't want to remove it as it needs to remain backwards compatible.

The code for Maintenance was updated, but since none of our tools yet use it, the logic is only in theory based on matching changes.

Additionally, I had no luck getting FTP stuff to work before hand.  I made changes until things started to operate.  I also tried to support using ssl:// like we did before, but this never worked for me, so I forced the use of the dedicated FTPS logic.  If anyone is able to get it to work, we can better identify the file system handlings.

Finally, I know there is leagues's file system (MIT License) handler that does FTP(s), SFTP, Webdav and a few others.  I thought about using it, but the only additional library I see useful for SMF to support would be webdav.  If we want to redirect to using it, then we can.