[3.0] change mysqli_report to be enabled during debugging
When mysqli_report first changed in 8.1, we disabled it i #7126

We have since improved our database standards to fix any issues that strict mode would generate.

This change sets it so when using db_show_debug, we enable the errors/strict mode.  This should ensure that during development processes, we do not introduce any new issues that are being silently ignored.

It would be nice to add `MYSQLI_REPORT_INDEX` to the list, but it seems we can't yet.  It reports no index for a basic `select * from smf_settings`, which while true is not something we can fix due to the table design.