[3.0] Overhaul User IP handling #8697
Fixes #7684

This PR addresses the above issue and isolates the handling of IP determination into a dedicated function, while also removing all instances where we modify $_SERVER['REMOTE_ADDR'] or set $_SERVER['BAN_CHECK_IP'], except when in compatibility mode in QueryString.php.  The latter purpose is to have SMF stop modifying variables on the server, leaving them as they were and cleaning them to its standards, and using them when needed.

This fixes the PR by reversing the order in which we perform the checks.  In the past, we have determined that the user is being passed through the reverse proxy and set that user's real IP address into BAN_CHECK_IP, leaving REMOTE_ADDR as the proxy.  We then proceeded to check and use REMOTE_ADDR most of the time.  By flipping this, we get the user's real IP and use that, and use the alternative IP as the original requesting IP (reverse proxy).

Reissuing the PR as #8697 is broken.