[3.0] Fix quotefast failures
The logic is greatly different than it was for 2.1.  Hopefully this resolves it.

On another note, when we get an error during the XML process, the ajax logic does not drop into error condition because a response was received (even for an error 500) and still tries to process it.  I did not try to address that here.  In all honesty updating our ajax logic to use json responses in my mind is when we ensure that is addressed.