[2.1] Fix <move-file>/<move-dir> in package-info
The test was trying to confirm that the destination file existed, when
it's actually supposed to not exist since we're moving something to
that name.

Turns out a similar fix was done earlier to stop the warning from
showing up for <create-dir> and <create-file> (#3009), so I just
extended that fix.

Fixes #8986

Signed-off-by: Dav999 <dav999.tolp@gmail.com>